Fixed issue where New File/Folder reported an error.
When creating a new folder where a folder named "New Folder" already existed, the extension reported an error instead of creating a new folder named "New folder (n)". This was due to an error in CGDriveShlExt::_NextNameSegment that did not return a pidl to a child element if it was called with a null pdwAttributes pointer.

Additionally, when parse was called with 0x2 as the flag, we were returning an error instead of the pidl. Correcting these errors allows the shell to continue to query with different names until it is able to create a new file with the correct number in parenthesis.

Increased version number.